+++ CHANGELOG.md
@@ -8,7 +8,9 @@ All notable changes to this project will be documented in this file.
 - Add ``get_slick_reporting_media`` and ``get_charts_media`` templatetags
 - Simplify adding export options and customizing the builtin export to csv button
 - Simplify adding custom buttons to the report page
-
+- Add `get_group_by_custom_querysets` hook to ReportView
+- Fix in SlickReportingListView
+ 
 ## [1.1.1] - 2023-09-25
 - Change settings to be a dict , adding support JQUERY_URL and FONT AWESOME customization #79 & #81
 - Fix issue with chartjs not being loaded  #80